separate file, RuntimeTypeHandlerEqualityComparere.cs
